refactor(app): #7 extracted OneTextField as util component
